requirements.txt
scipy
scikit-learn
h5py
tensorflow-gpu
